Refactor: 자유게시판경로 변경
