+++ styles/ArticleDetail.module.css
@@ -1,0 +1,160 @@
+.wrapper {
+    display: flex;
+    flex-direction: column;
+
+    margin-bottom: 36rem;
+
+    gap: 6.4rem;
+}
+
+.back-list {
+    display: flex;
+    align-items: center;
+
+    width: auto;
+    height: 4.8rem;
+
+    margin: 0 auto;
+    padding: 1.2rem 3.95rem;
+    border-radius: 4rem;
+
+    color: #F3F4F6;
+    font-size: 1.8rem;
+    font-weight: 600;
+    line-height: 2.6rem;
+
+    background-color: #3692FF;
+
+    gap: 0.8rem;
+}
+
+.article {
+    display: flex;
+    flex-direction: column;
+
+    gap: 2.4rem
+}
+
+.article p {
+    color: #1F2937;
+    font-size: 1.8rem;
+    font-weight: 400;
+    line-height: 2.6rem;
+}
+
+.article-header {
+    display: flex;
+    flex-direction: column;
+
+    padding-bottom: 1.6rem;
+    border-bottom: 0.1rem solid #E5E7EB;
+
+    gap: 1.6rem;
+}
+
+.article-title {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.article-title h1 {
+    font-size: 2rem;
+    font-weight: 700;
+    line-height: 3.2rem;
+}
+
+.menu {
+    position: relative;
+    cursor: pointer;
+}
+
+.dropdown {
+    position: absolute;
+    top: 70%;
+    right: 50%;
+}
+
+.info {
+    display: flex;
+    align-items: center;
+
+    gap: 3.2rem;
+}
+
+.user-info {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    gap: 1.6rem;
+}
+
+.user-info .image {
+    position: relative;
+}
+
+.user-info p {
+    font-size: 1.4rem;
+    line-height: 2.4rem;
+}
+
+.user-info .name {
+    color: #4B5563;
+    font-weight: 500;
+}
+
+.user-info .date {
+    color: #9CA3AF;
+    font-weight: 400;
+}
+
+.like-wrap {
+    border-left: 0.1rem solid #E5E7EB;
+}
+
+.like {
+    display: flex;
+    align-items: center;
+
+    height: 4rem;
+
+    margin-left: 3.2rem;
+    padding: 0.4rem 1.2rem;
+    border-radius: 3.5rem;
+    border: 0.1rem solid #E5E7EB;
+
+    gap: 1rem;
+}
+
+.like p {
+    color: #6B7280;
+    font-size: 1.6rem;
+    font-weight: 500;
+    line-height: 2.6rem;
+}
+
+/* -- Mobile -- */
+@media screen and (max-width:743px) {
+    .user-info>:nth-child(2) {
+        color: #4B5563;
+        font-size: 1.2rem;
+        font-weight: 500;
+        line-height: 2.4rem;
+    }
+
+    .user-info>:last-child {
+        color: #9CA3AF;
+        font-size: 1.2rem;
+        font-weight: 400;
+        line-height: 2.4rem;
+    }
+
+    .info {
+        gap: 1rem;
+    }
+
+    .like {
+        margin-left: 1rem;
+    }
+}
